Remove invalid UTF-8 characters from text output
Sometimes broadcasters put "strange" characters in DVB strings and
when the strings are converted to UTF-8 the output contains
invalid characters.

In order to avoid this problem, this commit instructs iconv
library to ignore characters that can not be converted to
UTF-8 by adding switching the default output charset from
UTF-8 to UTF-8//IGNORE to the output charset.